remove unnecessary git/mingw code
