+++ win32_pthread.h
@@ -11,6 +11,7 @@
 #define WIN32_LEAN_AND_MEAN
 #endif
 
+#include "platform.h"
 #include <windows.h>
 
 /*
